requirements.txt
fastapi>=0.115.12,<0.116.0
uvicorn>=0.34.0,<0.35.00
sqlalchemy>=2.0.40,<2.1.0
asyncpg>=0.30.0,<0.31.0
python-dotenv>=1.1.0,<1.2.0
psycopg2>=2.9.10,<2.10.0